Day 17.1,2 Conway cubes
Forgot to submit part 1 separately, but part 2 just added the W dimension to the record type, and the extra for loop to query the neighbors. The rest of the solution is literally the same.

Seeing this, it should be trivial to modify it for an arbitrary number of dimensions, although that means either hard compiling the exact number of dimensions, or figure out something to make a dynamic array (or a record containing a dynamic array) the key for a hash set.

I used the spring set here, but in hindsight, this could perfectly have been built with a bunch of stringlists, to keep it simpler.
Also, I have separate sets for adding and removing cubes. This could also have been solved by using a dictionary of cubes with a state. I started that way, but it felt more complicated to build.