Day 18.2 Expression tree
Rewrite Day 1 into an expression tree, based on parentheses
For Day 2, apply some post processing (Nest), to move parts of the expression to a sub-expression, as if it is surrounded in parentheses. This enforces operator precedence.
Advantage: Various versions of nesting could be written, to implement math of different universes.